Added loop, new timing functionality
The loop will cycle through all markets. This works, but unfortunately
it is not recording the results for each cycle...

More importantly, I've added a second wait feature that waits for the
"loading" image that appears between inputs to disappear  before
continuing on. This seems to be a little more reliable and quicker than
set sleep times and looking for the next element.